modules/panel-type-classifier/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "panel-type-classifier"
version = "0.1.0"
description = "Zero-shot panel-type classification (imaging vs non-imaging + subtype) using BiomedCLIP."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "open_clip_torch==2.23.0",
  "torch",
  "pillow",
  "tqdm",
  "pandas",
  "matplotlib",
  "numpy",
]

[tool.setuptools]
packages = ["panel_type_classifier"]
